requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-sock==0.7.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httptools==0.6.4
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.3.2
passlib==1.7.4
pdf2image==1.17.0
pillow==11.3.0
priority==2.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.14.1
PyJWT==2.10.1
pymongo==4.6.0
PyPDF2==3.0.1
python-dotenv==1.0.0
python-engineio==4.12.2
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.2
redis==5.0.1
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
taskgroup==0.2.2
tomli==2.2.1
typing_extensions==4.14.1
utils==1.0.2
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==12.0
Werkzeug==3.1.3
wsproto==1.2.0
